remove redundant methods
